requirements.txt
tree_sitter==0.20.1
tree_sitter_java==0.23.4
ollama==0.1.6
tqdm==4.66.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
pylint==3.0.2
pyyaml==6.0.1
loguru==0.7.2